+++ conjur/cli.py
@@ -23,6 +23,8 @@ from conjur.client import Client
 from conjur.constants import DEFAULT_NETRC_FILE, DEFAULT_CONFIG_FILE
 from conjur.credentials_data import CredentialsData
 from conjur.credentials_from_file import CredentialsFromFile
+from conjur.list import ListData, ListController
+from conjur.list.list_logic import ListLogic
 from conjur.login import LoginLogic, LoginController
 from conjur.logout import LogoutController, LogoutLogic
 from conjur.version import __version__
@@ -186,8 +188,28 @@ Copyright 2020 CyberArk Software Ltd. All rights reserved.
         # *************** LIST COMMAND ***************
 
         # pylint: disable=line-too-long
-        resource_subparsers.add_parser('list',
-                                       help='List all available resources belonging to this account')
+        list_subparser = resource_subparsers.add_parser('list',
+                                                         help='List all available resources belonging to this account')
+
+        list_options = list_subparser.add_argument_group(title=self.title("Options"))
+        list_options.add_argument('-i', '--inspect',
+                                  action='store_true', dest='inspect',
+                                  help='')
+        list_options.add_argument('-k', '--kind',
+                                  action='store', dest='kind',
+                                  help='')
+        list_options.add_argument('-r', '--role',
+                                  action='store', dest='role',
+                                  help='')
+        list_options.add_argument('-l', '--limit',
+                                  action='store', dest='limit',
+                                  help='')
+        list_options.add_argument('-o', '--offset',
+                                  action='store', dest='offset',
+                                  help='')
+        list_options.add_argument('-s', '--search',
+                                  action='store', dest='search',
+                                  help='')
 
         # *************** POLICY COMMAND ***************
 
@@ -296,16 +318,34 @@ Copyright 2020 CyberArk Software Ltd. All rights reserved.
         login_controller.load()
 
     @classmethod
-    def handle_logout_logic(cls):
+    def handle_logout_logic(cls, ssl_verify=True):
         """
         Method that wraps the logout call logic
         """
         credentials = CredentialsFromFile(DEFAULT_NETRC_FILE)
         logout_logic = LogoutLogic(credentials)
 
-        logout_controller = LogoutController(ssl_verify=None,
+        logout_controller = LogoutController(ssl_verify=ssl_verify,
                                              logout_logic=logout_logic)
         logout_controller.remove_credentials()
+
+    @classmethod
+    # pylint: disable=too-many-arguments
+    def handle_list_logic(cls, ssl_verify=True, kind=None,
+                          inspect=False, search=None,
+                          limit=None, offset=None,
+                          role=None, client=None):
+        """
+        Method that wraps the list call logic
+        """
+        list_data = ListData(kind=kind, inspect=inspect,
+                             search=search, limit=limit,
+                             offset=offset, acting_as=role)
+        list_logic = ListLogic(client)
+        list_controller = ListController(ssl_verify=ssl_verify,
+                                         list_logic=list_logic,
+                                         list_data=list_data)
+        list_controller.load()
 
     @staticmethod
     # pylint: disable=too-many-branches
@@ -325,7 +365,7 @@ Copyright 2020 CyberArk Software Ltd. All rights reserved.
             Cli.handle_login_logic(args.name,  args.password, args.ssl_verify)
             return
         elif resource == 'logout':
-            Cli.handle_logout_logic()
+            Cli.handle_logout_logic(args.ssl_verify)
             return
 
         # Needed for unit tests so that they do not require configuring
@@ -345,8 +385,10 @@ Copyright 2020 CyberArk Software Ltd. All rights reserved.
         client = Client(ssl_verify=args.ssl_verify, debug=args.debug)
 
         if resource == 'list':
-            result = client.list()
-            print(json.dumps(result, indent=4))
+            Cli.handle_list_logic(args.ssl_verify, args.kind,
+                                  args.inspect, args.search,
+                                  args.limit, args.offset,
+                                  args.role, client)
         elif resource == 'whoami':
             result = client.whoami()
             print(json.dumps(result, indent=4))